#8638 org.jkiss.dbeaver.ext.exasol.ui created, Activator, tools(maintenance), ConnectionPage moved from ext.exasol to ext.exasol.ui
