+++ plugins/org.jkiss.dbeaver.ext.exasol.ui/OSGI-INF/l10n/bundle.properties
@@ -1,0 +1,14 @@
+#Generated by ResourceBundle Editor (http://essiembre.github.io/eclipse-rbe/)
+# DBeaver - Universal Database Manager
+# Copyright (C) 2016 Karl Griesser (fullref@gmail.com)
+# Copyright (C) 2010-2020 DBeaver Corp and others serge@jkiss.org
+
+Bundle-Name = DBeaver Exasol UI
+Bundle-Vendor = JKISS
+
+dialog.connection.header = Exasol Connection Settings
+
+tool.org.jkiss.dbeaver.ext.exasol.table.export.description = Export Tables, Views or Schemas to CSV
+tool.org.jkiss.dbeaver.ext.exasol.table.export.name        = Export Table(s)
+tool.org.jkiss.dbeaver.ext.exasol.table.import.description = Import Tables or Schemas from CSV
+tool.org.jkiss.dbeaver.ext.exasol.table.import.name        = Import Table(s)
