+++ plugins/org.jkiss.dbeaver.ext.exasol.ui/OSGI-INF/l10n/bundle_de.properties
@@ -1,0 +1,10 @@
+#Generated by ResourceBundle Editor (http://essiembre.github.io/eclipse-rbe/)
+# DBeaver - Universal Database Manager
+# Copyright (C) 2010-2020 DBeaver Corp and others
+# Copyright (C) 2017 Karl Griesser (fullref@gmail.com)
+dialog.connection.header = Exasol Verbindungseinstellungen
+
+tool.org.jkiss.dbeaver.ext.exasol.table.export.description = Exportieren von Tabellen, Ansichten oder Schemas nach CSV
+tool.org.jkiss.dbeaver.ext.exasol.table.export.name        = Tabelle(n) exportieren
+tool.org.jkiss.dbeaver.ext.exasol.table.import.description = Tabellen oder Schemas aus CSV importieren
+tool.org.jkiss.dbeaver.ext.exasol.table.import.name        = Tabelle(n) importieren
