+++ plugins/org.jkiss.dbeaver.ext.exasol.ui/OSGI-INF/l10n/bundle_ja.properties
@@ -1,0 +1,11 @@
+#Generated by ResourceBundle Editor (http://essiembre.github.io/eclipse-rbe/)
+# DBeaver - Universal Database Manager
+# Copyright (C) 2016 Karl Griesser (fullref@gmail.com)
+# Copyright (C) 2010-2020 DBeaver Corp and others serge@jkiss.org
+
+dialog.connection.header = \u00A0Exasol\u306E\u63A5\u7D9A\u8A2D\u5B9A
+
+tool.org.jkiss.dbeaver.ext.exasol.table.export.description = \u00A0\u8868\u3001\u30D3\u30E5\u30FC\u307E\u305F\u306F\u30B9\u30AD\u30FC\u30DE\u3092CSV\u306B\u30A8\u30AF\u30B9\u30DD\u30FC\u30C8\u3059\u308B
+tool.org.jkiss.dbeaver.ext.exasol.table.export.name        = \u00A0\u8F38\u51FA\u8868
+tool.org.jkiss.dbeaver.ext.exasol.table.import.description = \u00A0CSV\u304B\u3089\u30C6\u30FC\u30D6\u30EB\u307E\u305F\u306F\u30B9\u30AD\u30FC\u30DE\u3092\u30A4\u30F3\u30DD\u30FC\u30C8\u3059\u308B
+tool.org.jkiss.dbeaver.ext.exasol.table.import.name        = \u00A0\u8868\u306E\u30A4\u30F3\u30DD\u30FC\u30C8
