+++ plugins/org.jkiss.dbeaver.ext.exasol.ui/OSGI-INF/l10n/bundle_pt_BR.properties
@@ -1,0 +1,14 @@
+#Generated by ResourceBundle Editor (http://essiembre.github.io/eclipse-rbe/)
+# DBeaver - Universal Database Manager
+# Copyright (C) 2016 Karl Griesser (fullref@gmail.com)
+# Copyright (C) 2010-2020 DBeaver Corp and others serge@jkiss.org
+
+Bundle-Name = DBeaver Exasol UI
+Bundle-Vendor = JKISS
+
+dialog.connection.header = Configura\u00E7\u00F5es de conex\u00E3o Exasol 
+
+tool.org.jkiss.dbeaver.ext.exasol.table.export.description = Exportar Tabelas, Views ou  Esquema para CSV
+tool.org.jkiss.dbeaver.ext.exasol.table.export.name        = Exportar Tabela(s)
+tool.org.jkiss.dbeaver.ext.exasol.table.import.description = Importar Tabelas ou a partir de  CSV
+tool.org.jkiss.dbeaver.ext.exasol.table.import.name        = Importar  Tabela(s)
