+++ plugins/org.jkiss.dbeaver.ext.exasol.ui/OSGI-INF/l10n/bundle_ru.properties
@@ -1,0 +1,7 @@
+
+dialog.connection.header = \u041F\u0430\u0440\u0430\u043C\u0435\u0442\u0440\u044B \u043F\u043E\u0434\u043A\u043B\u044E\u0447\u0435\u043D\u0438\u044F Exasol
+
+tool.org.jkiss.dbeaver.ext.exasol.table.export.description = \u042D\u043A\u0441\u043F\u043E\u0440\u0442 \u0442\u0430\u0431\u043B\u0438\u0446, \u043F\u0440\u0435\u0434\u0441\u0442\u0430\u0432\u043B\u0435\u043D\u0438\u0439 \u0438\u043B\u0438 \u0441\u0445\u0435\u043C \u0432 CSV
+tool.org.jkiss.dbeaver.ext.exasol.table.export.name        = \u042D\u043A\u0441\u043F\u043E\u0440\u0442 \u0442\u0430\u0431\u043B\u0438\u0446(\u044B)
+tool.org.jkiss.dbeaver.ext.exasol.table.import.description = \u0418\u043C\u043F\u043E\u0440\u0442\u0438\u0440\u043E\u0432\u0430\u0442\u044C \u0442\u0430\u0431\u043B\u0438\u0446\u044B \u0438\u043B\u0438 \u0441\u0445\u0435\u043C\u044B \u0438\u0437 CSV
+tool.org.jkiss.dbeaver.ext.exasol.table.import.name        = \u0418\u043C\u043F\u043E\u0440\u0442 \u0442\u0430\u0431\u043B\u0438\u0446(\u044B)
