+++ plugins/org.jkiss.dbeaver.ext.exasol.ui/OSGI-INF/l10n/bundle_zh.properties
@@ -1,0 +1,8 @@
+#Generated by ResourceBundle Editor (http://essiembre.github.io/eclipse-rbe/)
+# Copyright (C) 2017 Liu, Yuanyuan (liuyuanyuan@highgo.com)
+dialog.connection.header = Exasol \u8FDE\u63A5\u8BBE\u7F6E
+
+tool.org.jkiss.dbeaver.ext.exasol.table.export.description = \u5BFC\u51FA\u8868, \u89C6\u56FE\u6216\u8005\u6A21\u5F0F\u5230 CSV
+tool.org.jkiss.dbeaver.ext.exasol.table.export.name        = \u5BFC\u51FA\u8868
+tool.org.jkiss.dbeaver.ext.exasol.table.import.description = \u4ECE CSV \u5BFC\u5165\u8868\u6216\u6A21\u5F0F
+tool.org.jkiss.dbeaver.ext.exasol.table.import.name        = \u5BFC\u5165\u8868
