+++ plugins/org.jkiss.dbeaver.ext.exasol.ui/build.properties
@@ -1,0 +1,7 @@
+source.. = src/
+output.. = target/classes/
+bin.includes = plugin.xml,\
+               META-INF/,\
+               .,\
+               OSGI-INF/,\
+               icons/,\
