+++ plugins/org.jkiss.dbeaver.ext.exasol.ui/plugin.xml
@@ -1,0 +1,56 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<?eclipse version="3.2"?>
+
+<plugin>
+
+    <extension point="org.jkiss.dbeaver.dataSourceView">
+        <view
+                id="org.jkiss.dbeaver.ext.exasol.dataSourceWizardView"
+                dataSource="exasol"
+                targetID="org.jkiss.dbeaver.ext.ui.newConnectionWizard"
+                label="%dialog.connection.header"
+                class="org.jkiss.dbeaver.ext.exasol.ui.views.ExasolConnectionPage"
+                icon="platform:/plugin/org.jkiss.dbeaver.ext.exasol/icons/Exasol_icon.png">
+        </view>
+        <view
+                id="org.jkiss.dbeaver.ext.exasol.dataSourceEditorView"
+                dataSource="exasol"
+                targetID="org.jkiss.dbeaver.ext.ui.editConnectionDialog"
+                label="%dialog.connection.header"
+                class="org.jkiss.dbeaver.ext.exasol.ui.views.ExasolConnectionPage">
+        </view>
+    </extension>
+    
+        <extension point="org.jkiss.dbeaver.tools">
+      <tools>
+         <tool
+               class="org.jkiss.dbeaver.ext.exasol.ui.tools.ExasolExportTableTool"
+               description="%tool.org.jkiss.dbeaver.ext.exasol.table.export.description"
+               id="org.jkiss.dbeaver.ext.exasol.ui.tools.ExasolExportTableTool"
+               label="%tool.org.jkiss.dbeaver.ext.exasol.table.export.name">
+            <objectType
+                  name="org.jkiss.dbeaver.ext.exasol.model.ExasolTable">
+            </objectType>
+            <objectType
+                  name="org.jkiss.dbeaver.ext.exasol.model.ExasolView">
+            </objectType>
+            <objectType
+                  name="org.jkiss.dbeaver.ext.exasol.model.ExasolSchema">
+            </objectType>
+         </tool>
+         <tool
+               class="org.jkiss.dbeaver.ext.exasol.ui.tools.ExasolImportTableTool"
+               description="%tool.org.jkiss.dbeaver.ext.exasol.table.import.description"
+               id="org.jkiss.dbeaver.ext.exasol.ui.tools.ExasolImportTableTool"
+               label="%tool.org.jkiss.dbeaver.ext.exasol.table.import.name">
+            <objectType
+                  name="org.jkiss.dbeaver.ext.exasol.model.ExasolSchema">
+            </objectType>
+            <objectType
+                  name="org.jkiss.dbeaver.ext.exasol.model.ExasolTable">
+            </objectType>
+         </tool>
+      </tools>
+    </extension>
+
+</plugin>
