+++ plugins/org.jkiss.dbeaver.ext.exasol.ui/src/org/jkiss/dbeaver/ext/exasol/ui/internal/ExasolResources.properties
@@ -1,0 +1,29 @@
+dialog_connection_password = Password
+dialog_connection_port = Port
+dialog_connection_user_name = User name
+
+dialog_table_open_output_directory = Output Directory
+dialog_table_tools_column_heading = Column Headings?
+dialog_table_tools_column_sep = Column Separator
+dialog_table_tools_encoding = Encoding
+dialog_table_tools_export_compress = Compressed File?
+dialog_table_tools_export_title = Export to CSV File
+dialog_table_tools_file_template = File Template
+dialog_table_tools_import_title = Import from CSV File
+dialog_table_tools_options = Options
+dialog_table_tools_progress = Progress
+dialog_table_tools_result = Result
+dialog_table_tools_row_sep_mode = Use Row Separator
+dialog_table_tools_string_sep = String Separator
+dialog_table_tools_string_sep_mode = Use String Separator
+dialog_table_tools_success_title = Success
+
+exasol_partition_description = Information about columns that are used for partitioning
+exasol_partition_name = Partition info
+
+label_backup_host_list = Backup Host List
+label_database = Database
+label_encrypt = Encrypt Communication
+label_host_list = Host List
+label_security = Security
+label_use_backup_host_list = Use Backup Host List
