+++ plugins/org.jkiss.dbeaver.ext.exasol.ui/src/org/jkiss/dbeaver/ext/exasol/ui/internal/ExasolResources_de.properties
@@ -1,0 +1,34 @@
+#Generated by ResourceBundle Editor (http://essiembre.github.io/eclipse-rbe/)
+# DBeaver - Universal Database Manager
+# Copyright (C) 2010-2020 DBeaver Corp and others
+# Copyright (C) 2017 Karl Griesser (fullref@gmail.com)
+
+dialog_connection_password = Passwort
+dialog_connection_port = Port
+dialog_connection_user_name = Benutzername
+
+dialog_table_open_output_directory = Ausgabeverzeichnis
+dialog_table_tools_column_heading = Spalten\u00FCberschriften?
+dialog_table_tools_column_sep = Spaltentrennzeichen
+dialog_table_tools_encoding = Zeichenkodierung
+dialog_table_tools_export_compress = Dateien komprimieren?
+dialog_table_tools_export_title = Export auf CSV-Datei
+dialog_table_tools_file_template = Dateinamen-Vorlage
+dialog_table_tools_import_title = Aus CSV-Datei importieren
+dialog_table_tools_options = Optionen
+dialog_table_tools_progress = Fortschritt
+dialog_table_tools_result = Ergebnis
+dialog_table_tools_row_sep_mode = Zeilentrennzeichen Verwendungsart
+dialog_table_tools_string_sep = Zeichentrenneichen
+dialog_table_tools_string_sep_mode = Zeichentrennzeichen Verwendungsart
+dialog_table_tools_success_title = Erfolg
+
+exasol_partition_description = Informationen zu den Partitionierungsspalten
+exasol_partition_name = Partitionsinformation
+
+label_backup_host_list = Backup Host Liste
+label_database = Datenbank
+label_encrypt = Kommunikation verschl\u00FCsseln
+label_host_list = Host Liste
+label_security = Sicherheit
+label_use_backup_host_list = Backup Host Liste verwenden
