+++ plugins/org.jkiss.dbeaver.ext.exasol.ui/src/org/jkiss/dbeaver/ext/exasol/ui/internal/ExasolResources_ja.properties
@@ -1,0 +1,19 @@
+dialog_connection_password = \u30D1\u30B9\u30EF\u30FC\u30C9
+dialog_connection_port = \u30DD\u30FC\u30C8
+dialog_connection_user_name = \u30E6\u30FC\u30B6\u30FC\u540D
+
+dialog_table_open_output_directory = \u51FA\u529B\u30C7\u30A3\u30EC\u30AF\u30C8\u30EA
+dialog_table_tools_column_heading = \u5217\u898B\u51FA\u3057\uFF1F
+dialog_table_tools_column_sep = \u30AB\u30E9\u30E0\u30BB\u30D1\u30EC\u30FC\u30BF
+dialog_table_tools_encoding = \u30A8\u30F3\u30B3\u30FC\u30C7\u30A3\u30F3\u30B0
+dialog_table_tools_export_compress = \u5727\u7E2E\u30D5\u30A1\u30A4\u30EB\u3067\u3059\u304B\uFF1F
+dialog_table_tools_export_title = CSV\u30D5\u30A1\u30A4\u30EB\u306B\u30A8\u30AF\u30B9\u30DD\u30FC\u30C8
+dialog_table_tools_file_template = \u30D5\u30A1\u30A4\u30EB\u30C6\u30F3\u30D7\u30EC\u30FC\u30C8
+dialog_table_tools_import_title = CSV\u30D5\u30A1\u30A4\u30EB\u304B\u3089\u306E\u30A4\u30F3\u30DD\u30FC\u30C8
+dialog_table_tools_options = \u30AA\u30D7\u30B7\u30E7\u30F3
+dialog_table_tools_progress = \u9032\u6357
+dialog_table_tools_result = \u7D50\u679C
+dialog_table_tools_row_sep_mode = \u884C\u533A\u5207\u308A\u8A18\u53F7
+dialog_table_tools_string_sep = \u30B9\u30C8\u30EA\u30F3\u30B0\u30BB\u30D1\u30EC\u30FC\u30BF
+dialog_table_tools_string_sep_mode = \u6587\u5B57\u5217\u30BB\u30D1\u30EC\u30FC\u30BF\u3092\u4F7F\u7528\u3059\u308B
+dialog_table_tools_success_title = \u6210\u529F
