+++ plugins/org.jkiss.dbeaver.ext.exasol.ui/src/org/jkiss/dbeaver/ext/exasol/ui/internal/ExasolResources_ru.properties
@@ -1,0 +1,26 @@
+dialog_connection_password = \u041F\u0430\u0440\u043E\u043B\u044C
+dialog_connection_port = \u041F\u043E\u0440\u0442
+dialog_connection_user_name = \u0418\u043C\u044F \u043F\u043E\u043B\u044C\u0437\u043E\u0432\u0430\u0442\u0435\u043B\u044F
+
+dialog_table_open_output_directory = \u041A\u0430\u0442\u0430\u043B\u043E\u0433
+dialog_table_tools_column_heading = \u0417\u0430\u0433\u043E\u043B\u043E\u0432\u043A\u0438 \u0441\u0442\u043E\u043B\u0431\u0446\u043E\u0432?
+dialog_table_tools_column_sep = \u0420\u0430\u0437\u0434\u0435\u043B\u0438\u0442\u0435\u043B\u044C \u0441\u0442\u043E\u043B\u0431\u0446\u043E\u0432
+dialog_table_tools_encoding = \u041A\u043E\u0434\u0438\u0440\u043E\u0432\u043A\u0430
+dialog_table_tools_export_compress = \u0421\u0436\u0430\u0442\u044B\u0439 \u0444\u0430\u0439\u043B?
+dialog_table_tools_export_title = \u042D\u043A\u0441\u043F\u043E\u0440\u0442 \u0432 CSV \u0444\u0430\u0439\u043B
+dialog_table_tools_file_template = \u0428\u0430\u0431\u043B\u043E\u043D \u0444\u0430\u0439\u043B\u0430
+dialog_table_tools_import_title = \u0418\u043C\u043F\u043E\u0440\u0442 \u0438\u0437 CSV \u0444\u0430\u0439\u043B\u0430
+dialog_table_tools_options = \u041D\u0430\u0441\u0442\u0440\u043E\u0439\u043A\u0438
+dialog_table_tools_progress = \u041F\u0440\u043E\u0433\u0440\u0435\u0441\u0441
+dialog_table_tools_result = \u0420\u0435\u0437\u0443\u043B\u044C\u0442\u0430\u0442
+dialog_table_tools_row_sep_mode = \u0418\u0441\u043F\u043E\u043B\u044C\u0437\u043E\u0432\u0430\u0442\u044C Row Separator
+dialog_table_tools_string_sep_mode = \u0418\u0441\u043F\u043E\u043B\u044C\u0437\u043E\u0432\u0430\u0442\u044C String Separator
+dialog_table_tools_string_sep = \u0420\u0430\u0437\u0434\u0435\u043B\u0438\u0442\u0435\u043B\u044C \u0441\u0442\u0440\u043E\u043A
+dialog_table_tools_success_title = \u0423\u0441\u043F\u0435\u0448\u043D\u043E
+
+label_backup_host_list = \u0421\u043F\u0438\u0441\u043E\u043A \u0440\u0435\u0437\u0435\u0440\u0432\u043D\u044B\u0445 \u0445\u043E\u0441\u0442\u043E\u0432
+label_database = \u0411\u0430\u0437\u0430 \u0434\u0430\u043D\u043D\u044B\u0445
+label_encrypt = \u0428\u0438\u0444\u0440\u043E\u0432\u0430\u043D\u0438\u0435 \u0441\u0432\u044F\u0437\u0438
+label_host_list = \u0421\u043F\u0438\u0441\u043E\u043A \u0445\u043E\u0441\u0442\u043E\u0432
+label_security = \u0411\u0435\u0437\u043E\u043F\u0430\u0441\u043D\u043E\u0441\u0442\u044C
+label_use_backup_host_list = \u0418\u0441\u043F\u043E\u043B\u044C\u0437\u043E\u0432\u0430\u0442\u044C \u0441\u043F\u0438\u0441\u043E\u043A \u0440\u0435\u0437\u0435\u0440\u0432\u043D\u044B\u0445 \u0445\u043E\u0441\u0442\u043E\u0432
