+++ plugins/org.jkiss.dbeaver.ext.exasol.ui/src/org/jkiss/dbeaver/ext/exasol/ui/internal/ExasolResources_zh.properties
@@ -1,0 +1,22 @@
+#Generated by ResourceBundle Editor (http://essiembre.github.io/eclipse-rbe/)
+# Copyright (C) 2017 Liu, Yuanyuan (liuyuanyuan@highgo.com)
+
+dialog_connection_password = \u5BC6\u7801
+dialog_connection_port = \u7AEF\u53E3\u53F7
+dialog_connection_user_name = \u7528\u6237\u540D
+
+dialog_table_open_output_directory = \u8F93\u51FA\u76EE\u5F55
+dialog_table_tools_column_heading = \u5217\u6807\u9898\uFF1F
+dialog_table_tools_column_sep = \u5217\u5206\u9694\u7B26
+dialog_table_tools_encoding = \u7F16\u7801
+dialog_table_tools_export_compress = \u538B\u7F29\u6587\u4EF6\uFF1F
+dialog_table_tools_export_title = \u5BFC\u51FA\u5230 CSV \u6587\u4EF6
+dialog_table_tools_file_template = \u6587\u4EF6\u6A21\u677F
+dialog_table_tools_import_title = \u4ECE CSV \u6587\u4EF6\u5BFC\u5165
+dialog_table_tools_options = \u9009\u9879
+dialog_table_tools_progress = \u8FDB\u7A0B
+dialog_table_tools_result = \u7ED3\u679C
+dialog_table_tools_row_sep_mode = \u884C\u5206\u9694\u7B26
+dialog_table_tools_string_sep = \u5B57\u7B26\u4E32\u5206\u9694\u7B26
+dialog_table_tools_string_sep_mode = \u4F7F\u7528\u5B57\u7B26\u4E32\u5206\u9694\u7B26
+dialog_table_tools_success_title = \u6210\u529F
