+++ plugins/org.jkiss.dbeaver.ext.exasol.ui/src/org/jkiss/dbeaver/ext/exasol/ui/tools/ExasolExportTableTool.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jkiss.dbeaver.ext.exasol.tools.maintenance;
+package org.jkiss.dbeaver.ext.exasol.ui.tools;
 
 import org.eclipse.ui.IWorkbenchPart;
 import org.eclipse.ui.IWorkbenchWindow;
@@ -23,6 +23,7 @@ import org.jkiss.dbeaver.DBException;
 import org.jkiss.dbeaver.ext.exasol.model.ExasolSchema;
 import org.jkiss.dbeaver.ext.exasol.model.ExasolTable;
 import org.jkiss.dbeaver.ext.exasol.model.ExasolTableBase;
+import org.jkiss.dbeaver.ext.exasol.model.ExasolView;
 import org.jkiss.dbeaver.model.runtime.VoidProgressMonitor;
 import org.jkiss.dbeaver.model.struct.DBSObject;
 import org.jkiss.dbeaver.ui.tools.IUserInterfaceTool;
@@ -32,17 +33,19 @@ import java.util.Collection;
 import java.util.HashSet;
 import java.util.List;
 
-public class ExasolImportTableTool implements IUserInterfaceTool {
+public class ExasolExportTableTool implements IUserInterfaceTool {
 
-	public ExasolImportTableTool()
+
+	public ExasolExportTableTool()
 	{
 	}
-	
+
 	@Override
 	public void execute(IWorkbenchWindow window, IWorkbenchPart activePart,
 			Collection<DBSObject> objects) throws DBException
 	{
 		List<ExasolTable> tables = CommonUtils.filterCollection(objects, ExasolTable.class);
+		List<ExasolView> views = CommonUtils.filterCollection(objects, ExasolView.class);
 		List<ExasolSchema> schemas = CommonUtils.filterCollection(objects, ExasolSchema.class);
 		
 		//add tables for all Schemas but ignore views in schema
@@ -61,11 +64,17 @@ public class ExasolImportTableTool implements IUserInterfaceTool {
 			tableBaseObjects.add((ExasolTableBase) table);
 		}
 		
+		//add views
+		for(ExasolView view : views)
+		{
+			tableBaseObjects.add((ExasolTableBase) view);
+		}
 		
 		if (!tableBaseObjects.isEmpty()) {
-			ExasolImportTableToolDialog dialog = new ExasolImportTableToolDialog(activePart.getSite(), tableBaseObjects) ;
+			ExasolExportTableToolDialog dialog = new ExasolExportTableToolDialog(activePart.getSite(), tableBaseObjects);
 			dialog.open();
 		}
-
+		
 	}
+
 }
