+++ plugins/org.jkiss.dbeaver.ext.exasol/src/org/jkiss/dbeaver/ext/exasol/ExasolResources.properties
@@ -3,12 +3,6 @@ dialog_connection_alter_message = If a connection is altered that has saved cred
 
 dialog_connection_alter_title = Connection Alter
 
-dialog_connection_password = Password
-
-dialog_connection_port = Port
-
-dialog_connection_user_name = User name
-
 dialog_consumer_group_cpu_weight = CPU Weight
 
 dialog_consumer_group_group_limit = RAM Limit
@@ -55,36 +49,6 @@ dialog_schema_drop_title = Drop Schema Warning
 
 dialog_struct_edit_fk_label_fk_name = Foreign Key Name
 
-dialog_table_open_output_directory = Output Directory
-
-dialog_table_tools_column_heading = Column Headings?
-
-dialog_table_tools_column_sep = Column Separator
-
-dialog_table_tools_encoding = Encoding
-
-dialog_table_tools_export_compress = Compressed File?
-
-dialog_table_tools_export_title = Export to CSV File
-
-dialog_table_tools_file_template = File Template
-
-dialog_table_tools_import_title = Import from CSV File
-
-dialog_table_tools_options = Options
-
-dialog_table_tools_progress = Progress
-
-dialog_table_tools_result = Result
-
-dialog_table_tools_row_sep = Row Separator
-
-dialog_table_tools_string_sep = String Separator
-
-dialog_table_tools_string_sep_mode = Use String Separator
-
-dialog_table_tools_success_title = Success
-
 edit_exasol_constraint_manager_dialog_title = Create Primary Key
 
 editors_exasol_session_editor_action_kill = Kill Session
@@ -95,25 +59,9 @@ editors_exasol_session_editor_title_kill_session = Kill Session
 
 editors_exasol_session_editor_title_kill_session_statement = Terminate Statement
 
-exasol_partition_description = Information about columns that are used for partitioning
-
-exasol_partition_name = Partition info
-
 exasol_security_policy_description = System wide password security policy
 
 exasol_security_policy_name = System Security Policy
-
-label_backup_host_list = Backup Host List
-
-label_database = Database
-
-label_encrypt = Encrypt Communication
-
-label_host_list = Host List
-
-label_security = Security
-
-label_use_backup_host_list = Use Backup Host List
 
 manager_assign_priority_group = Assign priority group
