+++ plugins/org.jkiss.dbeaver.ext.exasol/src/org/jkiss/dbeaver/ext/exasol/ExasolResources_de.properties
@@ -3,12 +3,6 @@
 # Copyright (C) 2010-2020 DBeaver Corp and others
 # Copyright (C) 2017 Karl Griesser (fullref@gmail.com)
 
-dialog_connection_password = Passwort
-
-dialog_connection_port = Port
-
-dialog_connection_user_name = Benutzername
-
 dialog_consumer_group_cpu_weight = CPU Priorisierung
 
 dialog_consumer_group_group_limit = RAM Limit
@@ -49,36 +43,6 @@ dialog_priority_group_weight = Gewicht der Priorit\u00E4tsgruppe
 
 dialog_struct_edit_fk_label_fk_name = Fremschl\u00FCsselname
 
-dialog_table_open_output_directory = Ausgabeverzeichnis
-
-dialog_table_tools_column_heading = Spalten\u00FCberschriften?
-
-dialog_table_tools_column_sep = Spaltentrennzeichen
-
-dialog_table_tools_encoding = Zeichenkodierung
-
-dialog_table_tools_export_compress = Dateien komprimieren?
-
-dialog_table_tools_export_title = Export auf CSV-Datei
-
-dialog_table_tools_file_template = Dateinamen-Vorlage
-
-dialog_table_tools_import_title = Aus CSV-Datei importieren
-
-dialog_table_tools_options = Optionen
-
-dialog_table_tools_progress = Fortschritt
-
-dialog_table_tools_result = Ergebnis
-
-dialog_table_tools_row_sep = Zeilentrennzeichen
-
-dialog_table_tools_string_sep = Zeichentrenneichen
-
-dialog_table_tools_string_sep_mode = Zeichentrennzeichen Verwendungsart
-
-dialog_table_tools_success_title = Erfolg
-
 edit_exasol_constraint_manager_dialog_title = Prim\u00E4rschl\u00FCssel Anlegen
 
 editors_exasol_session_editor_action_kill = Verbindung terminieren
@@ -88,10 +52,6 @@ editors_exasol_session_editor_confirm_action = {0} {1}. Sind Sie sicher?
 editors_exasol_session_editor_title_kill_session = Verbindung terminieren
 
 editors_exasol_session_editor_title_kill_session_statement = Befehl terminieren
-
-exasol_partition_description = Informationen zu den Partitionierungsspalten
-
-exasol_partition_name = Partitionsinformation
 
 exasol_security_policy_description = Systemweite Passwortsicherheitsrichtlinie
