+++ plugins/org.jkiss.dbeaver.ext.exasol/src/org/jkiss/dbeaver/ext/exasol/ExasolResources_ja.properties
@@ -3,12 +3,6 @@ dialog_connection_alter_message = \u8A3C\u660E\u66F8\u3092\u4FDD\u5B58\u3057\u30
 
 dialog_connection_alter_title = \u63A5\u7D9A\u306E\u5909\u66F4
 
-dialog_connection_password = \u30D1\u30B9\u30EF\u30FC\u30C9
-
-dialog_connection_port = \u30DD\u30FC\u30C8
-
-dialog_connection_user_name = \u30E6\u30FC\u30B6\u30FC\u540D
-
 dialog_general_continue = \u6301\u7D9A\u3059\u308B
 
 dialog_schema_drop_message = \u30B9\u30AD\u30FC\u30DE\u5185\u306E\u65E2\u5B58\u306E\u30AA\u30D6\u30B8\u30A7\u30AF\u30C8\u306F\u3059\u3079\u3066\u524A\u9664\u3055\u308C\u307E\u3059\u3002\u3053\u308C\u306F\u3001\u5143\u306B\u623B\u3059\u3053\u3068\u306F\u3067\u304D\u307E\u305B\u3093\u3002
@@ -16,36 +10,6 @@ dialog_schema_drop_message = \u30B9\u30AD\u30FC\u30DE\u5185\u306E\u65E2\u5B58\u3
 dialog_schema_drop_title = \u30B9\u30AD\u30FC\u30DE\u8B66\u544A\u306E\u30C9\u30ED\u30C3\u30D7
 
 dialog_struct_edit_fk_label_fk_name = \u5916\u90E8\u30AD\u30FC\u540D
-
-dialog_table_open_output_directory = \u51FA\u529B\u30C7\u30A3\u30EC\u30AF\u30C8\u30EA
-
-dialog_table_tools_column_heading = \u5217\u898B\u51FA\u3057\uFF1F
-
-dialog_table_tools_column_sep = \u30AB\u30E9\u30E0\u30BB\u30D1\u30EC\u30FC\u30BF
-
-dialog_table_tools_encoding = \u30A8\u30F3\u30B3\u30FC\u30C7\u30A3\u30F3\u30B0
-
-dialog_table_tools_export_compress = \u5727\u7E2E\u30D5\u30A1\u30A4\u30EB\u3067\u3059\u304B\uFF1F
-
-dialog_table_tools_export_title = CSV\u30D5\u30A1\u30A4\u30EB\u306B\u30A8\u30AF\u30B9\u30DD\u30FC\u30C8
-
-dialog_table_tools_file_template = \u30D5\u30A1\u30A4\u30EB\u30C6\u30F3\u30D7\u30EC\u30FC\u30C8
-
-dialog_table_tools_import_title = CSV\u30D5\u30A1\u30A4\u30EB\u304B\u3089\u306E\u30A4\u30F3\u30DD\u30FC\u30C8
-
-dialog_table_tools_options = \u30AA\u30D7\u30B7\u30E7\u30F3
-
-dialog_table_tools_progress = \u9032\u6357
-
-dialog_table_tools_result = \u7D50\u679C
-
-dialog_table_tools_row_sep = \u884C\u533A\u5207\u308A\u8A18\u53F7
-
-dialog_table_tools_string_sep = \u30B9\u30C8\u30EA\u30F3\u30B0\u30BB\u30D1\u30EC\u30FC\u30BF
-
-dialog_table_tools_string_sep_mode = \u6587\u5B57\u5217\u30BB\u30D1\u30EC\u30FC\u30BF\u3092\u4F7F\u7528\u3059\u308B
-
-dialog_table_tools_success_title = \u6210\u529F
 
 edit_exasol_constraint_manager_dialog_title = \u4E3B\u30AD\u30FC\u306E\u4F5C\u6210
