+++ plugins/org.jkiss.dbeaver.ext.exasol/src/org/jkiss/dbeaver/ext/exasol/ExasolResources_ru.properties
@@ -1,20 +1,1 @@
-
-dialog_connection_password = \u041F\u0430\u0440\u043E\u043B\u044C
-
-dialog_connection_port = \u041F\u043E\u0440\u0442
-
-dialog_connection_user_name = \u0418\u043C\u044F \u043F\u043E\u043B\u044C\u0437\u043E\u0432\u0430\u0442\u0435\u043B\u044F
-
 exasol_security_policy_name = \u041F\u043E\u043B\u0438\u0442\u0438\u043A\u0430 \u0431\u0435\u0437\u043E\u043F\u0430\u0441\u043D\u043E\u0441\u0442\u0438 \u0441\u0438\u0441\u0442\u0435\u043C\u044B
-
-label_backup_host_list = \u0421\u043F\u0438\u0441\u043E\u043A \u0440\u0435\u0437\u0435\u0440\u0432\u043D\u044B\u0445 \u0445\u043E\u0441\u0442\u043E\u0432
-
-label_database = \u0411\u0430\u0437\u0430 \u0434\u0430\u043D\u043D\u044B\u0445
-
-label_encrypt = \u0428\u0438\u0444\u0440\u043E\u0432\u0430\u043D\u0438\u0435 \u0441\u0432\u044F\u0437\u0438
-
-label_host_list = \u0421\u043F\u0438\u0441\u043E\u043A \u0445\u043E\u0441\u0442\u043E\u0432
-
-label_security = \u0411\u0435\u0437\u043E\u043F\u0430\u0441\u043D\u043E\u0441\u0442\u044C
-
-label_use_backup_host_list = \u0418\u0441\u043F\u043E\u043B\u044C\u0437\u043E\u0432\u0430\u0442\u044C \u0441\u043F\u0438\u0441\u043E\u043A \u0440\u0435\u0437\u0435\u0440\u0432\u043D\u044B\u0445 \u0445\u043E\u0441\u0442\u043E\u0432
