+++ plugins/org.jkiss.dbeaver.ext.exasol/src/org/jkiss/dbeaver/ext/exasol/ExasolResources_zh.properties
@@ -5,12 +5,6 @@ dialog_connection_alter_message = \u5982\u679C\u4E00\u4E2A\u4FDD\u5B58\u6388\u67
 
 dialog_connection_alter_title = \u8FDE\u63A5\u53D8\u66F4
 
-dialog_connection_password = \u5BC6\u7801
-
-dialog_connection_port = \u7AEF\u53E3\u53F7
-
-dialog_connection_user_name = \u7528\u6237\u540D
-
 dialog_general_continue = \u7EE7\u7EED
 
 dialog_schema_drop_message = schema\u4E0B\u7684\u6240\u6709\u73B0\u5B58\u5BF9\u8C61\u5C06\u88AB\u5220\u9664.\u8BE5\u64CD\u4F5C\u65E0\u6CD5\u88AB\u64A4\u9500.
@@ -18,36 +12,6 @@ dialog_schema_drop_message = schema\u4E0B\u7684\u6240\u6709\u73B0\u5B58\u5BF9\u8
 dialog_schema_drop_title = Schema\u5220\u9664\u8B66\u544A
 
 dialog_struct_edit_fk_label_fk_name = \u5916\u952E\u540D\u79F0
-
-dialog_table_open_output_directory = \u8F93\u51FA\u76EE\u5F55
-
-dialog_table_tools_column_heading = \u5217\u6807\u9898\uFF1F
-
-dialog_table_tools_column_sep = \u5217\u5206\u9694\u7B26
-
-dialog_table_tools_encoding = \u7F16\u7801
-
-dialog_table_tools_export_compress = \u538B\u7F29\u6587\u4EF6\uFF1F
-
-dialog_table_tools_export_title = \u5BFC\u51FA\u5230 CSV \u6587\u4EF6
-
-dialog_table_tools_file_template = \u6587\u4EF6\u6A21\u677F
-
-dialog_table_tools_import_title = \u4ECE CSV \u6587\u4EF6\u5BFC\u5165
-
-dialog_table_tools_options = \u9009\u9879
-
-dialog_table_tools_progress = \u8FDB\u7A0B
-
-dialog_table_tools_result = \u7ED3\u679C
-
-dialog_table_tools_row_sep = \u884C\u5206\u9694\u7B26
-
-dialog_table_tools_string_sep = \u5B57\u7B26\u4E32\u5206\u9694\u7B26
-
-dialog_table_tools_string_sep_mode = \u4F7F\u7528\u5B57\u7B26\u4E32\u5206\u9694\u7B26
-
-dialog_table_tools_success_title = \u6210\u529F
 
 edit_exasol_constraint_manager_dialog_title = \u521B\u5EFA\u4E3B\u952E
